+++ src/inspect_ai/_view/www/src/@types/log.d.ts
@@ -567,6 +567,7 @@ export type Input3 = (
   | ChatMessageAssistant
   | ChatMessageTool
 )[];
+export type InputRefs = number[][] | null;
 export type Name9 = string;
 export type Description2 = string;
 export type Type18 = "object";
@@ -588,6 +589,8 @@ export type Response = {
 } | null;
 export type Error3 = boolean | null;
 export type Time1 = number | null;
+export type CallRefs = number[][] | null;
+export type CallKey = string | null;
 export type Completed1 = string | null;
 export type WorkingTime = number | null;
 export type Uuid6 = string | null;
@@ -845,6 +848,13 @@ export type TotalTime = number | null;
 export type WorkingTime3 = number | null;
 export type Uuid19 = string | null;
 export type ErrorRetries = EvalError[] | null;
+export type MessagePool = (
+  | ChatMessageSystem
+  | ChatMessageUser
+  | ChatMessageAssistant
+  | ChatMessageTool
+)[];
+export type CallPool = JsonValue[];
 export type Type27 =
   | "context"
   | "time"
@@ -1338,6 +1348,8 @@ export interface EvalSample {
   error: EvalError | null;
   error_retries: ErrorRetries;
   attachments: Attachments;
+  message_pool: MessagePool;
+  call_pool: CallPool;
   limit: EvalSampleLimit | null;
 }
 /**
@@ -1728,6 +1740,7 @@ export interface ModelEvent {
   model: Model4;
   role: Role4;
   input: Input3;
+  input_refs: InputRefs;
   tools: Tools1;
   tool_choice: ToolChoice;
   config: GenerateConfig;
@@ -1796,6 +1809,8 @@ export interface ModelCall {
   response: Response;
   error: Error3;
   time: Time1;
+  call_refs: CallRefs;
+  call_key: CallKey;
 }
 export interface Request {
   [k: string]: JsonValue;